# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Skynet API Documentation",
        "version": "1.0.0",
        "title": "Salesfactory Skynet"
    },
    "schemes": ["http"],
    "host": "localhost:5000",
    "basePath": "/api",
    "paths": {
        "/users": {
            "get": {
                "summary": "Get all users",
                "tags": ["users"],
                "description": "Returns a list of users. Can be filtered by the search query parameter.",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query to filter users",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 200
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/User"
                                    }
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Success"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new user",
                "tags": ["users"],
                "description": "Creates a new user.",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "user",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 200
                                },
                                "data": {
                                    "$ref": "#/definitions/User"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Success"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 400
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Bad Request"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "1"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "uid": {
                    "type": "string",
                    "example": "6cb69c71-3b9a-4a0f-8f1a-4d7d2f2e0b0e"
                },
                "avatar": {
                    "type": "string",
                    "example": "https://example.com/avatar.jpg"
                },
                "description": {
                    "type": "string",
                    "example": "User description"
                },
                "contact": {
                    "type": "string",
                    "example": "Contact information"
                },
                "active": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}